# perf: circuit evaluation with dfs + chore: rm paralllel feature