--- a/src/bgg/circuit/eval.rs
+++ b/src/bgg/circuit/eval.rs
@@ -13,10 +13,8 @@ pub trait Evaluable:
     + for<'a> Add<&'a Self, Output = Self>
     + for<'a> Sub<&'a Self, Output = Self>
     + for<'a> Mul<&'a Self, Output = Self>
-    + Send
-    + Sync
 {
-    type Params: Debug + Clone + Send + Sync;
+    type Params: Debug + Clone;
     fn rotate(&self, params: &Self::Params, shift: usize) -> Self;
     fn from_bits(params: &Self::Params, one: &Self, bits: &[bool]) -> Self;
 }
@@ -34,7 +32,7 @@ impl<P: Poly> Evaluable for P {
         let poly = Self::const_zero(params);
         let mut coeffs = poly.coeffs();
         let one_elem = <P::Elem as PolyElem>::one(&params.modulus());
-        for (i, bit) in bits.iter().enumerate() {
+        for (i, bit) in bits.into_iter().enumerate() {
             if *bit {
                 coeffs[i] = one_elem.clone();
             }