--- a/src/bgg/circuit/mod.rs
+++ b/src/bgg/circuit/mod.rs
@@ -2,15 +2,10 @@ pub mod eval;
 pub mod gate;
 pub mod serde;
 pub mod utils;
-use dashmap::DashMap;
 pub use eval::*;
 pub use gate::{PolyGate, PolyGateType};
 use itertools::Itertools;
-use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
-use std::{
-    collections::{BTreeMap, HashMap},
-    fmt::Debug,
-};
+use std::{collections::BTreeMap, fmt::Debug};
 pub use utils::*;
 
 use crate::utils::debug_mem;
@@ -199,41 +194,11 @@ impl PolyCircuit {
         gate_id
     }
 
-    /// Computes a levelized grouping of gate ids.
-    /// Input wires (keys 0..=num_input) are assigned level 0.
-    /// Each non‐input gate's level is defined as max(levels of its inputs) + 1.
-    pub fn compute_levels(&self) -> Vec<Vec<usize>> {
-        let mut gate_levels: HashMap<usize, usize> = HashMap::new();
-        let mut levels: Vec<Vec<usize>> = vec![];
-        for i in 0..=self.num_input {
-            gate_levels.insert(i, 0);
-        }
-        for (&gate_id, gate) in self.gates.iter() {
-            // skipping input gates
-            if gate_id <= self.num_input {
-                continue;
-            }
-            let level = gate
-                .input_gates
-                .iter()
-                .map(|id| *gate_levels.get(id).unwrap_or(&0))
-                .max()
-                .unwrap_or(0) +
-                1;
-            gate_levels.insert(gate_id, level);
-            if levels.len() <= level {
-                levels.resize(level + 1, vec![]);
-            }
-            levels[level].push(gate_id);
-        }
-        levels
-    }
-
     fn eval_gate<E: Evaluable>(
         &self,
         params: &E::Params,
         one: &E,
-        wires: &DashMap<usize, E>,
+        wires: &mut BTreeMap<usize, E>,
         gate: &PolyGate,
     ) {
         if !wires.contains_key(&gate.gate_id) {
@@ -319,19 +284,15 @@ impl PolyCircuit {
             assert_eq!(self.num_input(), inputs.len());
             assert_ne!(self.num_output(), 0);
         }
-        let wires = DashMap::new();
+        let mut wires: BTreeMap<usize, E> = BTreeMap::new();
         wires.insert(0, one.clone());
         for (idx, input) in inputs.iter().enumerate() {
             wires.insert(idx + 1, input.clone());
         }
-        let levels = self.compute_levels();
-        for level in levels.iter() {
-            // All gates in the same level can be processed in parallel.
-            level.par_iter().for_each(|&gate_id| {
-                let gate = self.gates.get(&gate_id).expect("gate not found").clone();
-                self.eval_gate(params, one, &wires, &gate);
-            });
-            debug_mem("Evaluated gate in parallel");
+        let output_gates = self.output_ids.iter().map(|id| self.gates[id].clone()).collect_vec();
+        for gate in output_gates.iter() {
+            self.eval_gate(params, one, &mut wires, gate);
+            debug_mem("Evaluated polynomial gate");
         }
 
         self.output_ids