- Topologically Sort and Iterate: basically return dependency order that ensures for any gate, all its dependencies (input gates) are processed before the gate itself
- with given dependency order, we no need perform recursive evaluation for prior gate like this (https://github.com/MachinaIO/diamond-io/blob/main/src/bgg/circuit/mod.rs#L228-L234) , so to reduce non necessary function call
- also removed `paralllel` feature as we not needed 

## time  - dummy param - mac m2

before
```
Time to obfuscate: 297.504950042s
Time for evaluation: 402.378013875s
Total time: 699.882963917s
```

after
```
Time to obfuscate: 285.641014709s
Time for evaluation: 385.333062416s
Total time: 670.974077125s
```